Complying with both SPDK 16.08 and DPDK 16.11

+++ include/nvfuse_io_manager.h
@@ -102,7 +102,7 @@ struct io_job{
     long offset; // in bytes
     long discard_range[2]; //discard
     size_t bytes; // requested bytes
-    int ret; //return value of pread or pwrite
+    size_t ret; //return value of pread or pwrite
     int req_type; // is read 
     char *buf;
     int complete;
